% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{adjust_loh}
\alias{adjust_loh}
\title{Trim LOH segments with respect to loss segments.}
\usage{
adjust_loh(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the cleaned segments, their copy
number and copy number types.
}
\description{
Trim LOH segments with respect to loss segments.
}
\details{
LOH segments completely contained within (or equal to)  a copy loss
segment are deleted.
LOH segments partially overlapping (on one end only) with a copy loss segment
 are trimmed to remove the overlap or split into several segments.
}
\examples{
segs.adj <- adjust_loh(segs.chas_example)
}
