% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntypes.R
\docType{data}
\name{cntype.amp}
\alias{cntype.amp}
\title{copy number type 'Amplification' (3 additional copies or more).}
\format{
character
}
\usage{
cntype.amp
}
\description{
copy number type 'Amplification' (3 additional copies or more).
}
\keyword{datasets}
