% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_doc.R
\docType{data}
\name{cntype.gain}
\alias{cntype.gain}
\title{copy number type 'Gain' (one or more additional copies).
In some contextes can be defined as the gain of 1-2 extra copies but not more.}
\format{
character
}
\source{
\code{cntype.gain <- 'Gain'}
}
\usage{
cntype.gain
}
\description{
copy number type 'Gain' (one or more additional copies).
In some contextes can be defined as the gain of 1-2 extra copies but not more.
}
\keyword{datasets}
