% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntypes.R
\docType{data}
\name{cntype.loh}
\alias{cntype.loh}
\title{copy number type 'LOH' (copy neutral loss of heterozygosity).}
\format{
character
}
\usage{
cntype.loh
}
\description{
copy number type 'LOH' (copy neutral loss of heterozygosity).
}
\keyword{datasets}
