% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntypes.R
\docType{data}
\name{cntype.loss}
\alias{cntype.loss}
\title{copy number type 'Loss' (loss of one or more copies; depends on gender for
the sexual chromosomes).}
\format{
character
}
\usage{
cntype.loss
}
\description{
copy number type 'Loss' (loss of one or more copies; depends on gender for
the sexual chromosomes).
}
\keyword{datasets}
