% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntypes.R
\docType{data}
\name{cntype.strongamp}
\alias{cntype.strongamp}
\title{copy number type 'Strong amplification' (8 additional copies or more).}
\format{
character
}
\usage{
cntype.strongamp
}
\description{
copy number type 'Strong amplification' (8 additional copies or more).
}
\keyword{datasets}
