% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_amp_segments}
\alias{get_amp_segments}
\title{Return all segments with an amplification (5 or more copies).}
\usage{
get_amp_segments(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the selected segments, their copy
number and copy number types.
}
\description{
Return all segments with an amplification (5 or more copies).
}
\examples{
segs.amp <- get_amp_segments(segs.chas_example)
}
