% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_cn_subtype}
\alias{get_cn_subtype}
\title{Get the detailed copy number types.}
\usage{
get_cn_subtype(segments, gender)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy number (field \code{cn}) and
their copy number types (field \code{cn.type}). The \code{cn.type} is expected to be either "Gain", "Loss"
or "LOH".}

\item{gender}{Character to indicate whether the sample is male ("M") or female ("F"). If \code{NULL}
then the field \code{cn.type} is set to \code{NA} for the chromosomes X and Y.}
}
\value{
A list of copy number types (one for each segment). Raises an error if the \code{cn.type},
\code{cn} and gender are not compatible.
}
\description{
Get the detailed copy number types.
}
\details{
The copy number subtypes are defined as follow:
  - "Gain": 1-2 extra copies
  - "Weak amplification": 3-7 extra copies
  - "Strong amplification": 8 or more extra copies
  - "Heterozygote loss": Loss of one copy out of two
  - "Homozygote loss": Loss of all copies
  - "LOH": copy-neutral loss of one parental allele
}
\examples{
subtypes <- get_cn_subtype(segs.chas_example, 'F')
}
