% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_gain_segments}
\alias{get_gain_segments}
\title{Return all segments with gain of copies.}
\usage{
get_gain_segments(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the selected segments, their copy
number and copy number types.
}
\description{
Return all segments with gain of copies.
}
\examples{
segs.gain <- get_gain_segments(segs.chas_example)
}
