% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_hetloss_segments}
\alias{get_hetloss_segments}
\title{Return all segments with heterozygous loss.}
\usage{
get_hetloss_segments(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the selected segments, their copy
number and copy number types.
}
\description{
Return all segments with heterozygous loss.
}
\examples{
segs.hetloss <- get_hetloss_segments(segs.chas_example)
}
