% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_homloss_segments}
\alias{get_homloss_segments}
\title{Return all segments with homozygous loss.}
\usage{
get_homloss_segments(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the selected segments, their copy
number and copy number types.
}
\description{
Return all segments with homozygous loss.
}
\examples{
get_homloss_segments <- get_homloss_segments(segs.chas_example)
}
