% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_loh_segments}
\alias{get_loh_segments}
\title{Return all segments of type LOH, independently of the copy number.}
\usage{
get_loh_segments(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the selected segments, their copy
number and copy number types.
}
\description{
Return all segments of type LOH, independently of the copy number.
}
\examples{
segs.loh <- get_loh_segments(segs.chas_example)
}
