% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_loss_segments}
\alias{get_loss_segments}
\title{Return all segments with loss of 1 or 2 copies.}
\usage{
get_loss_segments(segments)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy
number and copy number types.}
}
\value{
A \code{GRanges} object containing the selected segments, their copy
number and copy number types.
}
\description{
Return all segments with loss of 1 or 2 copies.
}
\examples{
segs.loh <- get_loh_segments(segs.chas_example)
}
