% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cn_segment}
\alias{is_cn_segment}
\title{Test if the object is a CNV segment}
\usage{
is_cn_segment(obj, raise_error = TRUE)
}
\arguments{
\item{obj}{Object to test}

\item{raise_error}{Boolean. If TRUE then raises an error if the test fails.}
}
\value{
Boolean. TRUE if the object is of class \code{GRanges} with fields
\code{cn} and \code{cn.type}.
}
\description{
Test if the object is a CNV segment
}
\examples{
library(GenomicRanges)
library(IRanges)
s1 <- GRanges(seqnames = '1p', ranges = IRanges(start = 1, end = 10),
              cn = 6, cn.type = cntype.gain, cn.subtype = cntype.weakamp)
is_cn_segment(s1)
s2 <- GRanges(seqnames = '1p', ranges = IRanges(start = 1, end = 10),
              cn = 6)
is_cn_segment(s2, raise_error = FALSE)
}
