% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{load_bed}
\alias{load_bed}
\title{Load a CNV file in a BED-like format.}
\usage{
load_bed(filename, gender)
}
\arguments{
\item{filename}{Path to the CNV file.}

\item{gender}{Character to indicate whether the sample is male ('M') or
female ('F').}
}
\value{
A \code{GRanges} object containing the segments, their copy number
and copy number types. If the
file contains twice the same segment or does not respect the format
specifications, then an error is raised.
}
\description{
Load a CNV file in a BED-like format.
}
\details{
The file is expected to contain the following columns: chromosome
name, segment start, segment end,
copy number. An fifth column with the copy number type is optional and if
present should contain the values
'Gain', 'Loss' or 'LOH'.
If the gender is not specified, then the segments on the sexual chromosomes
(X and Y) are dropped as the
copy number type cannot be determined for those.
}
\examples{
segs.filename <- system.file('extdata', 'cnv_example.bed',
  package = 'oncoscanR')
segs.cnv_example <- load_bed(segs.filename, 'F')
}
