% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_doc.R
\docType{data}
\name{oncoscan_na33.cov}
\alias{oncoscan_na33.cov}
\title{Arm coverage of the Oncoscan na33.r1 assay}
\format{A \code{GRanges} object containing the regions covered on each chromosome arm. Arms not covered by
the kit are absent from the object.}
\source{
Processing of the file OncoScan.na33.r1.annot.csv (obtained from
https://www.affymetrix.com/analysis/downloads/na33/genotyping/) by the function
\code{oncoscan_na33.cov <- get_oncoscan_coverage_from_probes()}.
}
\usage{
oncoscan_na33.cov
}
\description{
Arm coverage of the Oncoscan na33.r1 assay
}
\examples{
\dontrun{
 oncoscan_na33.cov
}
}
\keyword{datasets}
