% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ploidy}
\alias{ploidy}
\title{Compute the total number of chromosomes (ploidy).}
\usage{
ploidy(segments, kit.coverage)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy number and copy number types.}

\item{kit.coverage}{A \code{GRanges} object containing the regions covered on each chromosome arm.}
}
\value{
the total number of chromosomes
}
\description{
Compute the total number of chromosomes (ploidy).
}
\details{
For each chromosome get the maximum copy number variation (positive or negative) that cover 80%
of each arm of the chromosome. If an arm is not part of the kit coverage, then it is assumed that the other
arm is equally altered.
}
\examples{
ploidy(segs.chas_example, kit.coverage = oncoscan_na33.cov)
}
