% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_segments}
\alias{same_segments}
\title{Compare two CNV segments and test if equal}
\usage{
same_segments(x, y)
}
\arguments{
\item{x}{A \code{GRanges} object one segment, with optional fields \code{cn}, \code{cn.type} and
\code{cn.subtype}}

\item{y}{A \code{GRanges} object one segment, with optional fields \code{cn}, \code{cn.type} and
\code{cn.subtype}}
}
\value{
A boolean. TRUE if the segments' chromosome/arm, start/end and CN tags match.
}
\description{
Compare two CNV segments and test if equal
}
\examples{
library(GenomicRanges)
library(IRanges)
s1 <- GRanges(seqnames = '1p', ranges = IRanges(start = 1, end = 10),
              cn = 5, cn.type = cntype.gain, cn.subtype = cntype.weakamp)
s2 <- GRanges(seqnames = '1p', ranges = IRanges(start = 1, end = 10),
              cn = 6, cn.type = cntype.gain, cn.subtype = cntype.weakamp)
same_segments(s1,s2)
}
