% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_segmentsets}
\alias{same_segmentsets}
\title{Small function to test if two sets have the same CN segments}
\usage{
same_segmentsets(grA, grB)
}
\arguments{
\item{grA}{A \code{GRanges} object with fields \code{cn} and \code{cn.type}}

\item{grB}{A \code{GRanges} object with fields \code{cn} and \code{cn.type}}
}
\value{
TRUE if the two sets contain exactly the same CN segments
}
\description{
Small function to test if two sets have the same CN segments
}
