% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_doc.R
\docType{data}
\name{segs.chas_example}
\alias{segs.chas_example}
\title{Expected segments from loading the ChAS file 'chas_example.txt'.}
\format{
A \code{GRanges} object containing the segments, their copy number (field \code{cn}) and
their copy number types (field \code{cn.type}).
}
\source{
segs.filename <- system.file("extdata", "chas_example.txt", package = "oncoscanR")
mykit.cov <- get_oncoscan_coverage_from_probes()
segs.chas_example <- load_chas(segs.filename, kit.coverage = mykit.cov)
}
\usage{
segs.chas_example
}
\description{
Expected segments from loading the ChAS file 'chas_example.txt'.
}
\keyword{datasets}
