% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_oncoscan.R
\name{workflow_oncoscan.run}
\alias{workflow_oncoscan.run}
\title{Run the standard workflow for Oncoscan ChAS files.}
\usage{
workflow_oncoscan.run(chas.fn, gender)
}
\arguments{
\item{chas.fn}{Path to the text-export ChAS file}

\item{gender}{Gender of the sample (M or F)}
}
\value{
A list of lists with the following elements:
\code{armlevel = list(AMP= list of arms, GAIN= list of arms, LOSS= list of arms, LOH= list of arms),
scores = list(LST= number, LOH= number, TDplus= number, TD= number),
gender = gender as given by the parameter,
file = path of the ChAS file as given by the parameter)}
}
\description{
Run the standard workflow for Oncoscan ChAS files.
}
\details{
Identifies the globally altered arms (\>=90\% of arm altered), computes the HRD and
TD+ scores. The amplification is defined as a CN subtype \code{cntype.weakamp} or
\code{cntype.strongamp}. An arm is gained if of CN type \code{cntype.gain} unless the arm is
amplified.
}
\examples{
segs.filename <- system.file("extdata", "chas_example.txt", package = "oncoscanR")
workflow_oncoscan.run(segs.filename, "M")
}
