% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_doc.R
\docType{data}
\name{cntype.hetloss}
\alias{cntype.hetloss}
\title{copy number type "Heterozygous loss" (loss of one of the two copies).}
\format{
character
}
\source{
\code{cntype.hetloss <- "Heterozygous loss"}
}
\usage{
cntype.hetloss
}
\description{
copy number type "Heterozygous loss" (loss of one of the two copies).
}
\keyword{datasets}
