% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_doc.R
\docType{data}
\name{cntype.weakamp}
\alias{cntype.weakamp}
\title{copy number type "Weak amplification" (3 to 7 additional copies).}
\format{
character
}
\source{
\code{cntype.weakamp <- "Weak amplification"}
}
\usage{
cntype.weakamp
}
\description{
copy number type "Weak amplification" (3 to 7 additional copies).
}
\keyword{datasets}
