% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cntypes}
\alias{cntypes}
\title{Accepted types of CN for the segments
  - 'Gain': 1-2 extra copies
  - 'Weak amplification': 3-7 extra copies
  - 'Strong amplification': 8 or more extra copies
  - 'Heterozygote loss': Loss of one copy out of two
  - 'Homozygote loss': Loss of all copies
  - 'LOH': copy-neutral loss of one parental allele
  @source
  cntypes <- list(LOH='LOH', Gain='Gain', Loss='Loss')}
\format{
An object of class \code{list} of length 3.
}
\usage{
cntypes
}
\description{
@noRd
}
\keyword{datasets}
