% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncoscan_na33.cov}
\alias{oncoscan_na33.cov}
\title{GenomicRanges object of the chromosomal arms coverage for the oncoscan assay
(based on file extdata/Oncoscan.na33.r2.cov.processed.bed).}
\format{
A \code{GRanges} object containing the region covered on each
chromosome arm.
}
\source{
\code{oncoscan_na33.cov <- get_oncoscan_coverage_from_bed(
    system.file('extdata', 'Oncoscan.na33.r2.cov.processed.bed',
                package = 'oncoscanR'))}
}
\usage{
oncoscan_na33.cov
}
\description{
GenomicRanges object of the chromosomal arms coverage for the oncoscan assay
(based on file extdata/Oncoscan.na33.r2.cov.processed.bed).
}
\keyword{datasets}
