% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{score_loh}
\alias{score_loh}
\title{Compute the number HR deficiency-associated LOH regions.}
\usage{
score_loh(segments, kit.coverage, armlevel.loh, armlevel.hetloss)
}
\arguments{
\item{segments}{A \code{GRanges} object containing the segments, their copy number and copy number types.}

\item{kit.coverage}{A \code{GRanges} object containing the regions covered on each chromosome arm.}

\item{armlevel.loh}{A list of arms with global/arm-level LOH alteration.}

\item{armlevel.hetloss}{A list of arms with global/arm-level heterozygous losses.}
}
\value{
An integer representing the number of HRD-LOH regions.
}
\description{
Compute the number HR deficiency-associated LOH regions.
}
\details{
Procedure based on the paper from Abkevich et al., Br J Cancer 2012 (PMID: 23047548). All LOH
segments larger than 15Mb but excluding chromosome with a global LOH alteration (to compute with the
\code{armlevel_alt} function on LOH segments only). This score was linked to BRCA1/2-deficient tumors.
Note that the function will merge overlapping or neighbor LOH segments (at a distance of 1bp).

Of note the cn.subtype has to be defined first via the function \code{get_cn_subtype}.
}
\examples{
segs.chas_example$cn.subtype <- get_cn_subtype(segs.chas_example, 'F')
armlevel.loh <- armlevel_alt(segs.chas_example[segs.chas_example$cn.type == cntype.loh],
                             kit.coverage = oncoscan_na33.cov)
armlevel.hetloss <- armlevel_alt(segs.chas_example[segs.chas_example$cn.subtype == cntype.hetloss],
                             kit.coverage = oncoscan_na33.cov)
score_loh(segs.chas_example, oncoscan_na33.cov, names(armlevel.loh), names(armlevel.hetloss))
}
